@@ -131,7 +131,7 @@
     "@types/node": "^14",
     "@types/react": "^17",
     "@types/react-dom": "^17",
-    "react": "^17",
+    "react": "^18",
     "typescript": "^4",
     "jest": "^27",
     "@scripts/jest-config": "workspace:*"
