@@ -62,7 +62,7 @@
     "@loadable/server": "^5.15.1",
     "@modern-js/webpack": "workspace:^1.12.0",
     "@modern-js/utils": "workspace:^1.7.12",
-    "react-dom": "^17",
+    "react-dom": "^18.2.0",
     "serialize-javascript": "^6.0.0"
   },
   "devDependencies": {
@@ -84,11 +84,11 @@
     "@types/react-helmet": "^6.1.2",
     "@scripts/build": "workspace:*",
     "jest": "^27",
-    "react": "^17",
+    "react": "^18.2.0",
     "@scripts/jest-config": "workspace:*"
   },
   "peerDependencies": {
-    "react": "^17",
+    "react": "^18.2.0",
     "react-helmet": "^6.1.0",
     "@loadable/component": "^5.15.0",
     "@modern-js/runtime-core": "workspace:^1.5.3"
